requirements.txt
Flask>=2.2
tensorflow-cpu>=2.11
numpy
scikit-learn
Pillow
gunicorn   


